cdk/requirements.txt
attrs==20.3.0
boto3==1.17.64
botocore==1.20.64
cattrs==1.5.0
constructs==3.3.75
aws-cdk.core~=1.0
aws-cdk.aws_ssm~=1.0
jmespath==0.10.0
jsii==1.29.0
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.10.0.0
urllib3==1.26.5
